Implementing Authorization using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -76,6 +76,18 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        /*
+        if(\Gate::allows('update-question',$question))
+        {
+            return view("questions.edit", compact('question'));
+        }
+        abort(403, "Access Denied.");
+        */
+
+        if(\Gate::denies('update-question',$question))
+        {
+            abort(403, "Access Denied.");
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -88,6 +100,11 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question',$question))
+        {
+            abort(403, "Access Denied.");
+        }
+
         $question->update($request->only('title','body'));
 
         return redirect('/questions')->with('success', "Your question has been updated.");
@@ -101,6 +118,11 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question))
+        {
+            abort(403, "Access Denied.");
+        }
+
         $question->delete();
 
         return redirect('questions')->with('success', "Your question has been deleted.");
